backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cssselect2==0.8.0
ctranslate2==4.5.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gmssl==3.2.2
h11==0.14.0
hf-xet==1.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
humanize==4.12.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
kombu==5.5.2
lxml==5.4.0
macholib==1.16.3
Markdown==3.8
markdown-it-py==3.0.0
markdown_pdf==1.7
MarkupSafe==3.0.2
mdurl==0.1.2
modelscope==1.25.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.3
numpy==2.2.4
onnxruntime==1.21.0
openai==1.70.0
orjson==3.10.16
packaging==24.2
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.2
pydantic_core==2.33.1
pydyf==0.11.0
PyExecJS==1.5.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.2
PyMuPDF==1.25.3
pyphen==0.17.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.1
sympy==1.13.1
SQLAlchemy==2.0.41
psycopg2-binary==2.9.12
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0; sys_platform != "win32"
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
yarl==1.19.0
yt-dlp==2025.3.31
zopfli==0.2.3.post1
